requirements.txt
# Azure ML
azureml-sdk==1.48.0
azureml-core==1.48.0
azureml-automl-core==1.48.0

# Deep Learning
tensorflow==2.13.0
torch==2.0.1
pytorch-lightning==2.0.2
transformers==4.30.2
gluonts==0.13.2

# Time Series
prophet==1.1.4
statsmodels==0.14.0

# Web Framework
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
openpyxl==3.1.2

# Visualization
plotly==5.15.0

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Monitoring
opencensus-ext-azure==1.1.11
psutil==5.9.8
gputil==1.4.0
mlflow==2.10.0
